+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+### Version 1.0.23 Tag: v1.0.23 (2019-01-06)
+ * Moved video playback into fragment
+ * Added duration to video list
+ * Cleaned up full screen button and activation
+
+### Version 1.0.22 Tag: v1.0.22 (2019-01-05)
+ * Cleaned up like/dislike
+ * Added video download
+
+### Version 1.0.21 Tag: v1.0.21 (2019-01-05)
+ * Added more video meta data
+ * Very basic like and dislike functionality
+ * UI changes to video detail page
+ * Torrent stream fatal fix (@lishoujun)
+ * AR Strings update (@rex07)
+ * ZH Strings update (@lishoujun)
+ * RU Strings update (@ferhadnecef)
+ * Refacturing (@lishoujun)
+
 ### Version 1.0.20 Tag: v1.0.20 (2019-01-02)
  * Added basic login framework
  * AR Strings update (@rex07)